Implemented fetching and parsing pages
Implemented requesting and parsing a page for links. There are still a
few bugs to work out. The most prominent being SSL errors for some
links. I'd also like to implement the ability to select a new link at
random if the request for the currenty selected link does not return OK.

I tested with a few urls and was able to, for the most part,
successfully generate a list of links written to DFT.json.